removed page header class, renamed directory
